+++ pipelines/capture/cittati/flows.py
@@ -4,7 +4,7 @@ from pipelines.capture.cittati.constants import constants
 from pipelines.capture.cittati.tasks import create_gps_extractor
 from pipelines.capture.templates.flows import create_default_capture_flow
 from pipelines.constants import constants as smtr_constants
-from pipelines.utils.prefect import set_default_parameters
+from pipelines.schedules import create_hourly_cron
 
 CAPTURA_REGISTROS_CITTATI = create_default_capture_flow(
     flow_name="cittati: registros - captura",
@@ -12,8 +12,8 @@ CAPTURA_REGISTROS_CITTATI = create_default_capture_flow(
     create_extractor_task=create_gps_extractor,
     agent_label=smtr_constants.RJ_SMTR_AGENT_LABEL.value,
     generate_schedule=False,
+    recapture_schedule_cron=create_hourly_cron(),
 )
-set_default_parameters(CAPTURA_REGISTROS_CITTATI, {"recapture": True})
 
 CAPTURA_REALOCACAO_CITTATI = create_default_capture_flow(
     flow_name="cittati: realocacao - captura",
@@ -21,5 +21,5 @@ CAPTURA_REALOCACAO_CITTATI = create_default_capture_flow(
     create_extractor_task=create_gps_extractor,
     agent_label=smtr_constants.RJ_SMTR_AGENT_LABEL.value,
     generate_schedule=False,
+    recapture_schedule_cron=create_hourly_cron(),
 )
-set_default_parameters(CAPTURA_REALOCACAO_CITTATI, {"recapture": True})